Add foundations for text rendering, minus textures
Refactor the primitive extraction and processing to defer serialization
to the primitive buffer until after gathering all primitives, including
text ones. This is required to count the number of non-empty glyphs per
text primitive, and transform the single text primitive into N rectangle
ones. This also opens the door to further optimizations like sorting
primitives based on overlapping rects, to bubble up texts and other
textured items and increase batching. Ideally we'd use bindless, but
this is not supported by wgpu, so we still need to bach primitives per
texture.